switch to full monthly amounts
