+++ policyengine_us/tests/policy/baseline/gov/local/ca/la/dwp/ez_save/ca_la_ez_save.yaml
@@ -13,3 +13,20 @@
     ca_la_ez_save_eligible: true
   output:
     ca_la_ez_save: 10
+
+- name: Monthly capped ez save amount
+  period: 2022-01
+  input:
+    pre_subsidy_electricity_expense: 
+      2022: 1_000
+    ca_la_ez_save_eligible: true
+  output:
+    ca_la_ez_save: 8.17
+
+- name: Monthly ez save amount with monthly expenses
+  period: 2022-01
+  input:
+    pre_subsidy_electricity_expense: 96
+    ca_la_ez_save_eligible: true
+  output:
+    ca_la_ez_save: 8
