--- policyengine_us/variables/gov/local/ca/la/dwp/ez_save/ca_la_ez_save_amount.py
@@ -1,17 +1,0 @@
-from policyengine_us.model_api import *
-
-
-class ca_la_ez_save_amount(Variable):
-    value_type = float
-    entity = Household
-    definition_period = MONTH
-    label = "Los Angeles County EZ Save program monthly benefit amount"
-    defined_for = "ca_la_ez_save_eligible"
-
-    def formula(household, period, parameters):
-        electricity_expense = add(
-            household, period, ["pre_subsidy_electricity_expense"]
-        )
-        p = parameters(period).gov.local.ca.la.dwp.ez_save
-        uncapped_amount = p.amount
-        return min_(electricity_expense, uncapped_amount)
